pyproject.toml

[tool.poetry]
name = "learning-modules"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.6"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
pillow = "^11.1.0"
black = "^25.1.0"
djangorestframework-simplejwt = "^5.4.0"
coverage = "^7.6.12"
drf-yasg = "^1.21.9"
isort = "^6.0.1"
celery = "^5.4.0"
redis = "^5.2.1"
gevent = "^24.11.1"
django-celery-beat = "^2.7.0"
ipython = "^9.0.0"
pybloom-live = "^4.0.0"
django-filter = "^25.1"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.1.2"
mypy = "^1.15.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-django = "^4.11.1"
flake8 = "^7.2.0"
pytest-cov = "^6.2.1"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py"]
testpaths = ["lm/tests", "users/tests"]
django_find_project = false
pythonpath = ["."]

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.black]
line-length = 119
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py
)
'''

[tool.isort]
line_length = 119

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"